+++ src/js/core/drop.js
@@ -1,7 +1,7 @@
 import Container from '../mixin/container';
 import Position from '../mixin/position';
 import Togglable from '../mixin/togglable';
-import {addClass, append, apply, css, hasClass, includes, isTouch, matches, MouseTracker, offset, on, once, parent, pointerCancel, pointerDown, pointerEnter, pointerLeave, pointerUp, query, removeClass, toggleClass, within} from 'uikit-util';
+import {addClass, append, apply, attr, css, hasClass, includes, isTouch, matches, MouseTracker, offset, on, once, parent, pointerCancel, pointerDown, pointerEnter, pointerLeave, pointerUp, query, removeClass, toggleClass, within} from 'uikit-util';
 
 export let active;
 
@@ -62,7 +62,8 @@ export default {
             this.target = this.$create('toggle', query(this.toggle, this.$el), {
                 target: this.$el,
                 mode: this.mode
-            });
+            }).$el;
+            attr(this.target, 'aria-haspopup', true);
         }
 
     },